// src/app/models/compra.ts
export class compra {
    correlativo: number;
    idcompra: number;
    dateadd: string;
    datemod: string;
    idsucursal: number;
    iduseradd: number;
    idusermod: number;
    nombre: string;
    precio: number;
    total: number;
    cantidad: number;
    factura: string;
    recibo: string;
}