// src/app/models/historialCajaInv.ts
export class historialCajaInv {
    idhistorialcaja: number;
    iniciocaja: number;
    monedas: number;
    compras: number;
    faltante: number;
    sobrante: number;
    totaltarjetas: number;
    ventatotal: number;
    ventasucursal: number;
    agricola: number;
    credomatic: number;
    davivienda: number;
    efectivo: number;
    pagoenlinea: number;
    cuscatlan: number;
    qrtransferencias: number;
    reservas: number;
    efectivoentregar: number;
    idsucursal: number;
    iduseradd: number;
    dateadd: string;
    idusermod: number;
    datemod: string;
}